Fix signed/unsigned comparison

+++ src/main.cpp
@@ -153,7 +153,7 @@ int initialiseTestbench()
 		1.0,1.0,1.0,1.0 
 	};
 
-	for (int i = 0; i < materialDefs.size(); i++)
+	for (size_t i = 0; i < materialDefs.size(); i++)
 	{
 		// set material name from text file?
 		Material* m = renderer->makeMaterial();
